+++ components/org.wso2.carbon.identity.api.user.approval/org.wso2.carbon.identity.rest.api.user.approval.v2/src/main/java/org/wso2/carbon/identity/rest/api/user/approval/v2/core/factories/ApprovalEventServiceFactory.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.rest.api.user.approval.v2.core.factories;
+
+import org.wso2.carbon.identity.workflow.engine.ApprovalTaskService;
+import org.wso2.carbon.identity.workflow.engine.ApprovalTaskServiceImpl;
+
+/**
+ * Implementation of Approval Management REST API.
+ */
+public class ApprovalEventServiceFactory {
+
+    private static final ApprovalTaskService SERVICE;
+
+    static {
+        SERVICE = new ApprovalTaskServiceImpl();
+    }
+
+    public static ApprovalTaskService getApprovalEventService() {
+
+        return SERVICE;
+    }
+}
