+++ components/org.wso2.carbon.identity.api.user.approval/pom.xml
@@ -34,7 +34,7 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>org.wso2.carbon.identity.rest.api.user.approval.v1</module>
+        <module>org.wso2.carbon.identity.rest.api.user.approval.v2</module>
         <module>org.wso2.carbon.identity.api.user.approval.common</module>
     </modules>
 </project>
